kadai1

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view("tasks.index", [
-            "tasks" => $tasks,
-        ]);
+        return view('welcome', $data);
     }
 
     /**
@@ -45,15 +53,15 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            "status" => "required|max:10",
+            
             "content" => "required|max:191",
         ]);
-        
-        $task= new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
         return redirect("/");
     }
 
@@ -66,10 +74,28 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.show', [
+        if (\Auth::id() === $task->user_id) {
+            
+            return view('tasks.show', [
             'task' => $task,
         ]);
+    
+        } else {
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        return view('welcome', $data);           
+
+        }
     }
 
     /**
@@ -106,7 +132,7 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->save();
 
-        return redirect('/');        
+        return redirect("/");       
     }
 
     /**
@@ -117,9 +143,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
-        return redirect('/');        
+        $task = \App\Task::find($id);
+        
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
+    
+        return redirect("/");
     }
 }
